Move info & index related tools to their own internal file
