+++ fooof/tests/test_core_utils.py
@@ -59,33 +59,6 @@ def test_check_array_dim():
     out = check_array_dim(np.array([1, 2, 3]))
     assert out.shape == (0, 3)
 
-def test_get_obj_desc():
-
-    desc =  get_obj_desc()
-
-    tfm = FOOOF()
-    objs = dir(tfm)
-
-    # Test that everything in dict is a valid component of the fooof object
-    for ke, va in desc.items():
-        for it in va:
-            assert it in objs
-
-def test_get_data_indices():
-
-    indices_fixed = get_data_indices('fixed')
-    assert indices_fixed
-    for ke, va in indices_fixed.items():
-        if ke == 'knee':
-            assert not va
-        else:
-            assert isinstance(va, int)
-
-    indices_knee = get_data_indices('knee')
-    assert indices_knee
-    for ke, va in indices_knee.items():
-        assert isinstance(va, int)
-
 def test_check_iter():
 
     # Note: generator case not tested
