+++ fooof/tests/test_fit.py
@@ -14,7 +14,8 @@ import pkg_resources as pkg
 from fooof import FOOOF
 from fooof.data import FOOOFSettings, FOOOFDataInfo, FOOOFResults
 from fooof.synth import gen_power_spectrum
-from fooof.core.utils import group_three, get_obj_desc
+from fooof.core.utils import group_three
+from fooof.core.info import get_obj_desc
 
 from fooof.tests.utils import get_tfm, plot_test
