+++ fooof/utils.py
@@ -3,7 +3,7 @@
 import numpy as np
 
 from fooof.synth import gen_freqs
-from fooof.core.utils import get_obj_desc
+from fooof.core.info import get_obj_desc
 
 ###################################################################################################
 ###################################################################################################
